+++ src/main/java/fi/dy/masa/malilib/util/EquipmentUtils.java
@@ -9,6 +9,7 @@ import net.minecraft.entity.EntityType;
 import net.minecraft.entity.EquipmentSlot;
 import net.minecraft.entity.attribute.EntityAttributes;
 import net.minecraft.item.*;
+import net.minecraft.registry.Registries;
 import net.minecraft.registry.RegistryKey;
 import net.minecraft.registry.entry.RegistryEntry;
 import net.minecraft.registry.tag.ItemTags;
@@ -49,7 +50,21 @@ public class EquipmentUtils
 			return false;
 		}
 
-		return stack.has(DataComponentTypes.WEAPON) || stack.isIn(ItemTags.WEAPON_ENCHANTABLE);
+		return (stack.isIn(ItemTags.WEAPON_ENCHANTABLE) ||
+			   isSword(stack) || isAxe(stack)) &&
+			   stack.has(DataComponentTypes.WEAPON);
+	}
+
+	public static boolean isSword(ItemStack stack)
+	{
+		String itemType = Registries.ITEM.getId(stack.getItem()).getPath();
+        return itemType.contains("_sword");
+    }
+
+	public static boolean isAxe(ItemStack stack)
+	{
+		String itemType = Registries.ITEM.getId(stack.getItem()).getPath();
+		return itemType.contains("_axe");
 	}
 
 	public static boolean isRangedWeapon(ItemStack stack)
@@ -85,7 +100,28 @@ public class EquipmentUtils
 			return false;
 		}
 
-		return stack.has(DataComponentTypes.TOOL) || stack.isIn(ItemTags.MINING_ENCHANTABLE);
+		return (stack.isIn(ItemTags.MINING_ENCHANTABLE) ||
+			   isPickAxe(stack) || isAxe(stack) ||
+			   isHoe(stack) || isShovel(stack)) &&
+			   stack.has(DataComponentTypes.TOOL);
+	}
+
+	public static boolean isPickAxe(ItemStack stack)
+	{
+		String itemType = Registries.ITEM.getId(stack.getItem()).getPath();
+		return itemType.contains("_pickaxe");
+	}
+
+	public static boolean isShovel(ItemStack stack)
+	{
+		String itemType = Registries.ITEM.getId(stack.getItem()).getPath();
+		return itemType.contains("_shovel");
+	}
+
+	public static boolean isHoe(ItemStack stack)
+	{
+		String itemType = Registries.ITEM.getId(stack.getItem()).getPath();
+		return itemType.contains("_hoe");
 	}
 
 	public static boolean isMiscTool(ItemStack stack)