+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## Changelog
-- sync `sakura-ryoko/malilib` 1.21.5-0.24.0-sakura.9
-- update NeoForge version to `21.5.47-beta`
+- sync `sakura-ryoko/malilib` 1.21.5-0.24.1
+- tag namespace `mafglib` -> `malilib`
+- update NeoForge version to `21.5.66-beta`